tweak solr_search_params over-ride
Still using the approach of over-riding solr_search_params instead of the new method of adding a logic block. The old approach still works. But eliminate trying
to deal with the old weird semantics of extra_controller_params, that doesn't ex
ist anymore and we don't have to do it. Use proper method signature.