when facet initially hidden, render flot on show
Flot can only render in a div with width and height, which
a hidden div with auto width/height doens't usually have. Previously,
we tried to briefly show and then re-hide the div to get a width
to render flot. That was proving hard to maintain, especially when
at smallest size in bootstrap5 there are multiple layers of collapse
involved.

Now, we only render flot if facet content is visible -- and catch
the bootstrap collapsible show event to render when div is shown,
if initially hidden.