Requires BL 5.14+ and runs without deprecation
I followed cbeer's lead
projectblacklight/blacklight_advanced_search@ca1b13acb165d05
for some integration. That is, a module included in CatalogController,
on inclusion will include a module in the configured SearchBuilder.
This is a bit wacky, but I wasn't sure the right safe way to do this
for all the current configuration possibilities, so just followed
cbeer's lead.

Note also at
https://github.com/projectblacklight/blacklight_range_limit/blob/bl514/lib/blacklight_range_limit/controller_override.rb#L41-L45
some contortions needed to execute a search_results intended to be
"default processor chain, but with one method removed, and another one
added", and do it in a safe way for various configurations. This may
be something BL API should support better built-in, with arguments
to SearchBuilder/CatalogController#search_results ?