Re-use tooltip instead of destroy/create
Previous destroy/create was making the tooltip change
not work at all for me in Chrome, with this error in js console:

~~~
tooltip.self-05afb17….js?body=1:236 Uncaught TypeError: Cannot read property 'trigger' of null
    at HTMLDivElement.complete (tooltip.self-05afb17….js?body=1:236)
    at HTMLDivElement.fn (jquery.self-bd7ddd3….js?body=1:4826)
    at HTMLDivElement.handle (transition.self-6ad2488….js?body=1:55)
    at HTMLDivElement.dispatch (jquery.self-bd7ddd3….js?body=1:5227)
    at HTMLDivElement.elemData.handle (jquery.self-bd7ddd3….js?body=1:4879)
    at Object.trigger (jquery.self-bd7ddd3….js?body=1:5131)
    at HTMLDivElement.<anonymous> (jquery.self-bd7ddd3….js?body=1:5861)
    at Function.each (jquery.self-bd7ddd3….js?body=1:371)
    at jQuery.fn.init.each (jquery.self-bd7ddd3….js?body=1:138)
    at jQuery.fn.init.trigger (jquery.self-bd7ddd3….js?body=1:5860)
~~~

I didn't quite get to the bottom of the change -- I think it was maybe
trying to add a new tooltip before the first one was actually fully removed,
or something like that.

Switching to-reusing one tooltip object but changing it's title works
properly again.

Changing title of an existing Bootstrap tooltip is not really documented,
but seems to be established and known to do it this way, and is using
BS tooltip public methods to do it. Eg
http://stackoverflow.com/questions/9501921/change-twitter-bootstrap-tooltip-content-on-click