Fix range facet fails in advanced form.
When in advanced form the form selector does not work because the range
is not rendered as a form within a form.  Instead it is a div.

This updates the selector to try again if form selection failed the
first time around with a selector that works for the facet when it is
inside of the advanced form.