remove explicit sprockets 'require jquery'
blacklight_range_limit code _does_ need JQuery to work (for it's own code and it's dependencies like flot).

It was probably on that basis that an explicit 'require jquery' was included in blacklight_range_limits sprockets-processed file to be included in the host app. While the host app may likely _also_ have a 'require jquery' in it's application.js, the assumption was that sprokets is smart enough not to require it twice (which _may_ be true, unsure).

However, 'require jquery' from jquery-rails is actually the (quite old) latest JQuery 1.x. The explicit require here precludes the host app from choosing to use the newer jquery 2.x or 3.x (which should probably work fine). If the host app _does_ try to 'require jquery3' to get jquery 3.x, it leads to a really confusing and hard to debug failure mode where two versions of JQuery are loaded on top of each other, and interfere with each other in odd ways.

By removing the explicit require, we can leave it up to the host app which version of JQuery to load. We can also let the host app decide to load jQuery from some _other_ place than the jquery-rails gem, such as a CDN, or via webpacker/yarn/npm.

This does mean that if an existing is not explicitly loading JQuery at all, and just counting on blacklight_range_limit to do it -- it would have to be changed to 'require jquery' in its application.js or otherwise load JQuery. This seems like a pretty unlikely situation though; among other things, the Blacklight generator adds a 'require jquery' to the host apps application.js already.

Closes #99