Remove unnecessary dummy_range field.
Apparently nine years ago when this change was added to blacklight,
blacklight required a search_field to be added or it would not show
results.  However this is no longer the case as one can verify by doing
a search in the demo site at https://demo.projectblacklight.org/?q=test.

Creating this dummy is causing issues on some of the blacklight
instances when a search is made without adding search_field.

See example here:
* https://librarysearch.temple.edu/catalog?utf8=%E2%9C%93&q=test&search_field=dummy_range&range%5Bpub_date_sort%5D%5Bbegin%5D=1504&range%5Bpub_date_sort%5D%5Bend%5D=2020&commit=Apply
* https://catalog.princeton.edu/?utf8=%E2%9C%93&q=test&search_field=dummy_range&range%5Bpub_date_start_sort%5D%5Bbegin%5D=1000&range%5Bpub_date_start_sort%5D%5Bend%5D=9999&commit=Limit

Note the "Dummy Range" label in the constraint widget.

This commit removes adding the dummy_range field when seach_field is not
defined as it is no longer required and causes issues on some current
blacklight instances.