Refactor JavaScript to enable adopter flexibility and customization
Inspired by some of the analysis and comments in #132, this is a
straight refactor to move JavaScript around. This allows for adopters to
modify BlacklightRangeLimit functions, without having to override entire
files.

For example:

```javascript
BlacklightRangeLimit.display_ratio = 1
```

Would produce a square plot area. Previously this type of customization
was not possible.