CSS file is not .scss and does not use any .scss features, should not end in .scss
This file only uses a sprockets 'require'. It does not use any scss features, and never has, since it's introduction in 8397f56112

I think it was a mistake initially to name it .css.scss. As we start exploring differnet ways to manage assets in the brave new world, it ending in .scss makes things more confusing, implying things about it that are not true -- that it needs to be @imported in an SCSS pipeline, or that it CAN be imported in an .scss pipeline (it won't work, as it the sprockets require won't be triggered if you do it that way!)

This should not disrupt anything, as the require apps use is just `require  'blacklight_range_limit'` anyway, doesn't actually have the .scss on it (and actually wouldn't work if it did), rails has just been somehow covering up for our mistake all along, but will work fine with it fixed.