For open-ended limit, ensure both limit and data min/max are taken account of in calculating range bucket boundaries
In our app we use a multi-valued field for our range limit. In that case, when entering an open-ended limit, range limit could accidentally calculate boundaries for the buckets that exceeded the specified single-end boundary. Because it was ignoring the boundaries unless both ends were present.

Fixed that logic, with tests.