hacky way to keep unpermitted params out of text facet links
It seems like there should be a better way -- I thought Blacklight search_state was only including allow-listed params in the first place but apparently not?  Not sure what a better way would be or if it would require Blacklight architectural changes, but this works for now. Regression test included that was red without this change.