Test that out of bounds ranges do not result in error
One test applying to formerly without tests #294.

In process, I realize that the values displayed in the "constraint" aren't clamped, so that'd be an area of improvement -- but I'm not too worried about it, I don't personally have a use case for it. The main thing is ensuring this bad data doesn't result in an uncaught exception, which this regression test now does.